Call driver functions through prototypes
All driver function pointers are stored as generic function pointers in
an array for ease of access, but calling conventions can differ across
platforms and calling through a generic function pointer can lead to
ABI issues.

This in particular affects ARM64 macOS, which differs from the standard
ARM64 ABI by not requiring arguments passed on the stack to be aligned
to an 8 byte boundary [1]. Specifically, this affects SQLSpecialColumns
as it has many arguments causing both the scope and nullable parameters
to be passed on the stack and both of these parameters are smaller than
8 bytes.

Now that the macro wrappers call through prototyped functions, various
warnings emerged. Most of them are related to SQLSet*Attr and
SQLSet*Option, which cast integers to pointers and vice-versa. There
were some more legitimate issues found passing pointers to integers of
the wrong size.

[1] https://developer.apple.com/documentation/xcode/writing_arm64_code_for_apple_platforms#//apple_ref/doc/uid/TP40013702-SW1